pyproject.toml
[tool.poetry]
name = "Sentiment Analysis Project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.18,<3.12"
numpy = "^1.26.4"
pandas = "^2.2.2"
tqdm = "^4.66.4"
gensim = "^4.3.3"
PyYAML = "^6.0.1"
nltk = "^3.8.1"
scikit-learn = "^1.5.1"
lightgbm = "^4.5.0"
tensorflow = "^2.17.0"
tensorflow-io-gcs-filesystem = "^0.31.0"

[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"